PostgreSQL tests kind of working
