+++ ydb/library/yql/providers/generic/connector/tests/datasource/postgresql/test.py
@@ -1,47 +1,66 @@
-from pathlib import Path
 import pytest
 
 from ydb.library.yql.providers.generic.connector.api.common.data_source_pb2 import EDataSourceKind
+from ydb.library.yql.providers.generic.connector.tests.utils.settings import Settings
+from ydb.library.yql.providers.generic.connector.tests.utils.runner import Runner
+import ydb.library.yql.providers.generic.connector.tests.utils.dqrun as dqrun
+import ydb.library.yql.providers.generic.connector.tests.utils.kqprun as kqprun
 
-from utils.settings import Settings
-import clickhouse
-import join
-import postgresql
-from test_cases.collection import Collection
-import test_cases.join
-import test_cases.select_missing_database
-import test_cases.select_missing_table
-import test_cases.select_positive_common
-import utils.clickhouse
-from utils.runner import Runner
-from conftest import configure_runner
-import utils.dqrun as dqrun
-import utils.kqprun as kqprun
-import utils.postgresql
+from client import Client
+from conftest import configure_runner, docker_compose_dir
+import scenario
+from collection import Collection
+
+import ydb.library.yql.providers.generic.connector.tests.test_cases.select_missing_database as select_missing_database
+import ydb.library.yql.providers.generic.connector.tests.test_cases.select_missing_table as select_missing_table
+import ydb.library.yql.providers.generic.connector.tests.test_cases.select_positive_common as select_positive_common
 
 
 # Global collection of test cases dependent on environment
-tc_collection = Collection(Settings.from_env())
+tc_collection = Collection(
+    Settings.from_env(docker_compose_dir=docker_compose_dir, data_source_kind=EDataSourceKind.POSTGRESQL)
+)
 
 runners = (dqrun.DqRunner, kqprun.KqpRunner)
 runners_ids = ("dqrun", "kqprun")
 
 
+# @pytest.mark.parametrize("runner_type", runners, ids=runners_ids)
+# @pytest.mark.parametrize("test_case", tc_collection.get('select_positive'), ids=tc_collection.ids('select_positive'))
+# @pytest.mark.usefixtures("settings")
+# @pytest.mark.usefixtures("postgresql_client")
+# def test_select_positive(
+#     request: pytest.FixtureRequest,
+#     settings: Settings,
+#     runner_type: Runner,
+#     postgresql_client: Client,
+#     test_case: select_positive_common.TestCase,
+# ):
+#     runner = configure_runner(runner=runner_type, settings=settings)
+#     scenario.select_positive(
+#         settings=settings,
+#         runner=runner,
+#         client=postgresql_client,
+#         test_case=test_case,
+#         test_name=request.node.name,
+#     )
+
+
 @pytest.mark.parametrize("runner_type", runners, ids=runners_ids)
 @pytest.mark.parametrize(
-    "test_case", tc_collection.get('select_positive_postgresql'), ids=tc_collection.ids('select_positive_postgresql')
+    "test_case", tc_collection.get('select_missing_database'), ids=tc_collection.ids('select_missing_database')
 )
 @pytest.mark.usefixtures("settings")
 @pytest.mark.usefixtures("postgresql_client")
-def test_select_positive_postgresql(
+def test_select_missing_database(
     request: pytest.FixtureRequest,
     settings: Settings,
     runner_type: Runner,
-    postgresql_client: utils.postgresql.Client,
-    test_case: test_cases.select_positive_common.TestCase,
+    postgresql_client: Client,
+    test_case: select_missing_database.TestCase,
 ):
     runner = configure_runner(runner=runner_type, settings=settings)
-    postgresql.select_positive(
+    scenario.select_missing_table(
         settings=settings,
         runner=runner,
         client=postgresql_client,
@@ -52,119 +71,23 @@ def test_select_positive_postgresql(
 
 @pytest.mark.parametrize("runner_type", runners, ids=runners_ids)
 @pytest.mark.parametrize(
-    "test_case", tc_collection.get('select_positive'), ids=tc_collection.ids('select_positive')
-)
-@pytest.mark.usefixtures("settings")
-@pytest.mark.usefixtures("clickhouse_client")
-def test_select_positive(
-    request: pytest.FixtureRequest,
-    settings: Settings,
-    runner_type: Runner,
-    clickhouse_client: utils.clickhouse.Client,
-    test_case: test_cases.select_positive_common.TestCase,
-):
-    runner = configure_runner(runner=runner_type, settings=settings)
-    clickhouse.select_positive(
-        test_name=request.node.name, settings=settings, runner=runner, client=clickhouse_client, test_case=test_case
-    )
-
-
-@pytest.mark.parametrize("runner_type", runners, ids=runners_ids)
-@pytest.mark.parametrize(
-    "test_case", tc_collection.get('select_missing_database'), ids=tc_collection.ids('select_missing_database')
-)
-@pytest.mark.usefixtures("settings")
-@pytest.mark.usefixtures("clickhouse_client")
-@pytest.mark.usefixtures("postgresql_client")
-def test_select_missing_database(
-    request: pytest.FixtureRequest,
-    tmp_path: Path,
-    settings: Settings,
-    runner_type: Runner,
-    clickhouse_client: utils.clickhouse.Client,
-    postgresql_client: utils.postgresql.Client,
-    test_case: test_cases.select_missing_database.TestCase,
-):
-    runner = configure_runner(runner=runner_type, settings=settings)
-    match test_case.data_source_kind:
-        case EDataSourceKind.CLICKHOUSE:
-            clickhouse.select_missing_table(
-                settings=settings,
-                runner=runner,
-                client=clickhouse_client,
-                test_case=test_case,
-                test_name=request.node.name,
-            )
-        case EDataSourceKind.POSTGRESQL:
-            postgresql.select_missing_table(
-                settings=settings,
-                runner=runner,
-                client=postgresql_client,
-                test_case=test_case,
-                test_name=request.node.name,
-            )
-        case _:
-            raise Exception(f'invalid data source: {test_case.data_source_kind}')
-
-
-@pytest.mark.parametrize("runner_type", runners, ids=runners_ids)
-@pytest.mark.parametrize(
     "test_case", tc_collection.get('select_missing_table'), ids=tc_collection.ids('select_missing_table')
 )
 @pytest.mark.usefixtures("settings")
-@pytest.mark.usefixtures("clickhouse_client")
 @pytest.mark.usefixtures("postgresql_client")
 def test_select_missing_table(
     request: pytest.FixtureRequest,
-    tmp_path: Path,
     settings: Settings,
     runner_type: Runner,
-    clickhouse_client: utils.clickhouse.Client,
-    postgresql_client: utils.postgresql.Client,
-    test_case: test_cases.select_missing_table.TestCase,
+    postgresql_client: Client,
+    test_case: select_missing_table.TestCase,
 ):
     runner = configure_runner(runner=runner_type, settings=settings)
-    match test_case.data_source_kind:
-        case EDataSourceKind.CLICKHOUSE:
-            clickhouse.select_missing_table(
-                test_name=request.node.name,
-                settings=settings,
-                runner=runner,
-                client=clickhouse_client,
-                test_case=test_case,
-            )
-        case EDataSourceKind.POSTGRESQL:
-            postgresql.select_missing_table(
-                test_name=request.node.name,
-                settings=settings,
-                runner=runner,
-                client=postgresql_client,
-                test_case=test_case,
-            )
-        case _:
-            raise Exception(f'invalid data source: {test_case.data_source_kind}')
-
-
-@pytest.mark.parametrize("runner_type", runners, ids=runners_ids)
-@pytest.mark.parametrize("test_case", tc_collection.get('join'), ids=tc_collection.ids('join'))
-@pytest.mark.usefixtures("settings")
-@pytest.mark.usefixtures("clickhouse_client")
-@pytest.mark.usefixtures("postgresql_client")
-def test_join(
-    request: pytest.FixtureRequest,
-    settings: Settings,
-    runner_type: Runner,
-    clickhouse_client: utils.clickhouse.Client,
-    postgresql_client: utils.postgresql.Client,
-    test_case: test_cases.join.TestCase,
-):
-    runner = configure_runner(runner=runner_type, settings=settings)
-    join.join(
+    scenario.select_missing_table(
         test_name=request.node.name,
-        clickhouse_client=clickhouse_client,
-        postgresql_client=postgresql_client,
-        runner=runner,
         settings=settings,
+        runner=runner,
+        client=postgresql_client,
         test_case=test_case,
     )
 
@@ -172,56 +95,41 @@ def test_join(
 @pytest.mark.parametrize("runner_type", runners, ids=runners_ids)
 @pytest.mark.parametrize("test_case", tc_collection.get('select_datetime'), ids=tc_collection.ids('select_datetime'))
 @pytest.mark.usefixtures("settings")
-@pytest.mark.usefixtures("clickhouse_client")
 @pytest.mark.usefixtures("postgresql_client")
 def test_select_datetime(
     request: pytest.FixtureRequest,
-    tmp_path: Path,
     settings: Settings,
     runner_type: Runner,
-    clickhouse_client: utils.clickhouse.Client,
-    postgresql_client: utils.postgresql.Client,
-    test_case: test_cases.select_positive_common.TestCase,
+    postgresql_client: Client,
+    test_case: select_positive_common.TestCase,
 ):
     runner = configure_runner(runner=runner_type, settings=settings)
-    match test_case.data_source_kind:
-        case EDataSourceKind.CLICKHOUSE:
-            clickhouse.select_positive(
-                test_name=request.node.name,
-                test_case=test_case,
-                settings=settings,
-                runner=runner,
-                client=clickhouse_client,
-            )
-        case EDataSourceKind.POSTGRESQL:
-            postgresql.select_positive(
-                settings=settings,
-                runner=runner,
-                client=postgresql_client,
-                test_case=test_case,
-                test_name=request.node.name,
-            )
-        case _:
-            raise Exception(f'invalid data source: {test_case.data_source_kind}')
+    scenario.select_positive(
+        settings=settings,
+        runner=runner,
+        client=postgresql_client,
+        test_case=test_case,
+        test_name=request.node.name,
+    )
 
 
 @pytest.mark.parametrize("runner_type", runners, ids=runners_ids)
 @pytest.mark.parametrize(
     "test_case",
-    tc_collection.get('select_positive_postgresql_schema'),
-    ids=tc_collection.ids('select_positive_postgresql_schema'),
+    tc_collection.get('select_positive_with_schema'),
+    ids=tc_collection.ids('select_positive_with_schema'),
 )
 @pytest.mark.usefixtures("settings")
 @pytest.mark.usefixtures("postgresql_client")
-def test_select_pg_schema(
+def test_select_positive_with_schema(
     request: pytest.FixtureRequest,
     settings: Settings,
     runner_type: Runner,
-    postgresql_client: utils.postgresql.Client,
-    test_case: test_cases.select_positive_common.TestCase,
+    postgresql_client: Client,
+    test_case: select_positive_common.TestCase,
 ):
     runner = configure_runner(runner=runner_type, settings=settings)
-    postgresql.select_pg_schema(
+    scenario.select_pg_schema(
         settings=settings,
         runner=runner,
         client=postgresql_client,
